server/requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.3.2
cycler==0.12.1
dash==4.0.0
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.133.1
fastjsonschema==2.21.2
Flask==3.1.3
flatbuffers==25.12.19
fonttools==4.61.1
h11==0.16.0
httptools==0.7.1
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
ipython==8.38.0
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mediapipe==0.10.8
mpmath==1.3.0
narwhals==2.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.2.6
onnxruntime==1.23.2
open3d==0.19.0
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.90
packaging==26.0
parso==0.8.6
pillow==12.1.1
platformdirs==4.9.2
plotly==6.5.2
prompt_toolkit==3.0.52
protobuf==3.20.3
pure_eval==0.2.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pyreadline3==3.5.4
# REALSENSE_REMOVE_CANDIDATE: pyrealsense2 RealSense SDK Python bindings
# - 웹캡 전환 시 제거 가능
pyrealsense2==2.56.5.9235  # REALSENSE_REMOVE_CANDIDATE
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
retrying==1.4.2
rpds-py==0.30.0
six==1.17.0
sounddevice==0.5.5
scipy==1.15.3
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
Werkzeug==3.1.6
widgetsnbextension==4.0.15
zipp==3.23.0
chumpy==0.70
